==> src/1.SAE_get_activations.ipynb <==
# 0. Set up / Installation / Basics

!nvidia-smi

!ceph_quota

NEURONPEDIA_KEY = "sk-np-BwxFua0jEx2cNSqsZPVlqmsfPgDKi47oEo7HAWXxiU00"
GEMMA2B_KEY = "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"

import os
PATH = "/users/k24086575/inf_narrative_msc/k24086575"
os.environ["HF_HOME"] = PATH  # e.g., $SCRATCH/hf_models
os.environ["TRANSFORMERS_CACHE"] = PATH

from huggingface_hub import login
import torch
GEMMA2B_KEY = "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"
login(token=GEMMA2B_KEY)
torch.set_grad_enabled(False)

# 1. Load/Save Dataset 

import pandas as pd

def load_dataset(path):
    df = pd.read_csv(path, encoding="utf-8-sig", quoting=1)
    return df

def save_dataset(path):
    small_df.to_csv(path, encoding="utf-8-sig", quoting=1, index=False)
pd.set_option('display.max_columns', None)

DATASET_PATH = "./programming.csv"
df = load_dataset(DATASET_PATH)

df.head()

df.describe()

df.info()

df.loc[df['type']=='programming', 'category'].unique()

# 2. Model + SAE Set up

import gc
gc.collect() # Python's garbage collector
torch.cuda.empty_cache()

from sae_lens import SAE
import transformer_lens

def load_model_and_sae(model_name, release, sae_id, device="cuda"):
    # Load model
    model = transformer_lens.HookedTransformer.from_pretrained(
                                                                model_name, 
                                                                device=device,
                                                                # device_map="auto",          # if using multiple gpu will be distributed
                                                                # torch_dtype=torch.float16   # half precision? 
                                                                # load_in_4bit=True         # 4 bit? 
    )
    # Load SAE for the given hook point
    sae = SAE.from_pretrained(release, sae_id)[0]
    sae = sae.to(device)
    # hook 
    hook_point = sae.cfg.hook_name

    # TODO: multi gpu loading for bigger models 
    # # ============================================================
    # # 2. Load the SAE to the CPU first to determine the hook point.
    # sae = SAE.from_pretrained(release, sae_id, device="cpu")[0]
    # hook_point = sae.cfg.hook_name
    # print(f"SAE hook point identified: {hook_point}")

    # # 3. Find the actual device of that hook point in the distributed model.
    # #    The hook point name corresponds to a module in the model.
    # hook_device = model.mod_dict[hook_point].device
    # print(f"Hook point is on device: {hook_device}")

    # # 4. Move the SAE to that specific device.
    # sae = sae.to(hook_device)
    # print(f"SAE moved to {sae.device}.")
    # # ============================================================
    return model, sae, hook_point


# GEMMA 2 2B
SAE_TYPE = "att"    # res/ mlp/ att
MODEL_NAME = "gemma-2-2b"
RELEASE = f"gemma-scope-2b-pt-{SAE_TYPE}-canonical"
LAYER = 25
SAE_ID = f"layer_{LAYER}/width_16k/canonical"
model, sae, hook_point = load_model_and_sae(model_name=MODEL_NAME, release=RELEASE, sae_id=SAE_ID, device="cuda")
# print("MODEL: ", model)
# print("SAE: ", sae)
print("HOOKPOINT: ", hook_point)


# DEEPSEEK
# deepseek-ai/DeepSeek-R1-Distill-Llama-8B
# from sae_lens import SAE
# release = "deepseek-r1-distill-llama-8b-qresearch"
# sae_id = "blocks.19.hook_resid_post"
# sae = SAE.from_pretrained(release, sae_id)[0]
# model_name = "deepseek-ai/DeepSeek-R1-Distill-Llama-8B"

from collections import Counter
from sae_lens.analysis.neuronpedia_integration import get_neuronpedia_quick_list

def analyze_prompt_with_sae(prompt, model, sae, hook_point, topk_mean=10, topk_token=3):
    with torch.no_grad():   # dont save info for backp
        hook_point = sae.cfg.hook_name
        # Run model with cache
        _, cache = model.run_with_cache(prompt, prepend_bos=True)
    
        # Encode activations from residuals using SAE
        residual_activations = cache[hook_point] 
        feature_acts = sae.encode(residual_activations)  # output : [batch_size, sequence_length, d_sae]
    
        # l0 = (feature_acts[:, 1:] > 0).float().sum(-1).detach()

        # Get top-k features by mean activation across sequence, for each separate token 
        # Using dim=0 gives you the top features for each separate token.
        # Using dim=1 gives you the single set of top features based on their average strength across the entire sentence.
        mean_acts = feature_acts.mean(dim=0)  
        top_vals, top_indices = torch.topk(mean_acts, topk_mean)
        top_mean_vals = top_vals.tolist()
        top_mean_ids = top_indices.tolist()
    
        # Get top-k features for each token (flattened)
        top_token_feats = torch.topk(feature_acts.squeeze(), topk_token).indices  # shape: [seq_len, topk]
        token_feature_ids = torch.flatten(top_token_feats).tolist()
    
        # reduce memory from flaot to int 
        for i in range(len(top_mean_vals)):
            for j in range(len(top_mean_vals[i])):
                top_mean_vals[i][j] = round(top_mean_vals[i][j], 2) 

        del cache, residual_activations, feature_acts
        gc.collect() # Python's garbage collector
        torch.cuda.empty_cache() # PyTorch's cache cleaner

        return top_mean_ids, top_mean_vals, token_feature_ids


prompt = """
87	Let’s remember the problem Mountain holidays (MOU1H). The story of Mountain holidays is the following. Some time ago, Chef discovered that more and more people have started climbing mountains every day. So he decided to build a restaurant in the Ukrainian resort Bukovel (Carpathian Mountains). But there are many places to choose, so he wants to choose the best one.
By your great helps, Chef could find to the best place to build a restaurant, now Chef becomes rich. Now he wants to build a restaurant in the next mountain. Similar to previous one, the next mountain is described by a sequence of N points. Here the points are numbered from 1 to N, and the height of point k is denoted by Hk. Every two adjacent points of the mountain are connected by a segment.


For example, the mountain described by N = 9, H = [0, 2, 1, 2, 1, 3, 0, 1, 0]  is like a following:


       /\
  /\/\/  \
 /        \/\

 In the mountain, all of the tourists will go from the point 1 to point N.


For comfort of the tourists, Chef has bought teleports too. Using a teleport, a tourist can be transferred from point i to point j, for all i < j. Of course, tourists can also walk from point i to point i+1 on foot.


Now Chef  wants to calculate the attractiveness of the mountain, as the number of the different climbs. 
A climb is defined by the nonempty sequence (p1, p1+1), (p2, p2+1), ..., (ps, ps+1) of the moves  on foot , where pk+1 ≤ pk+1 for k = 1, 2, ..., s − 1.

Two climbs, say (p1, p1+1), (p2, p2+1), ..., (ps, ps+1) and (q1, q1+1), (q2, q2+1), ..., (qt, qt+1) are different if and only if

s ≠ t or
There exists at least one k such that 1 ≤ k < min(s, t) and Hpk+1 – Hpk ≠ Hqk+1 – Hqk.

You are given the array H, find the number of the different climbs that exist on the mountain. Since the answer can be very large, output it modulo 10^9+9 = 1000000009.

Input
The first line of input contains T, denoting the number of test cases. Then T test cases follow.
The first line of each test case contains an integer N, denoting the number of points on the mountain.
On second line of each test case, there are N space-separated integers H1, H2, ..., HN, denoting the height of each point.

Output
For each test case, output an integer denoting the number of the different climbs modulo 10^9+9 = 1000000009.

Constraints

1 ≤ T ≤ 100000, that is, 1 ≤ T ≤ 10^5
2 ≤ N ≤ 1000000, that is, 2 ≤ N ≤ 10^6
Sum of N over all test cases in one file will be at most 1000000 = 10^6
-2000000 ≤ Hk ≤ 2000000, that is, −2 × 10^6 ≤ Hk ≤ 2 × 10^6

"""


top_mean_ids, top_mean_vals, token_feature_ids = analyze_prompt_with_sae(
    prompt=prompt,
    model=model,
    sae=sae,
    hook_point=hook_point,
    topk_mean=10,
    topk_token=3,
)

# # print(prompt)
# # print()
# # print("Top Activated Feature IDs:", top_mean_ids)
# # print()
# # print("Top Mean Activation Values:", top_mean_vals)
# # print()
# # print("All Token-wise Feature IDs:", token_feature_ids)
# # print()
print(get_neuronpedia_quick_list(sae=sae, features=token_feature_ids[:30], name=prompt))

import gc
gc.collect() # Python's garbage collector
torch.cuda.empty_cache()

def pipeline_process_datasets(df, DATASET_PATH, OUTPUT_PATH, model, sae, hook_point, release, layer):
    # top_mean_ids, top_mean_vals, token_feature_ids
    # all_top_mean_ids = []
    # all_top_mean_vals = []
    all_token_feature_ids = []
    for prompt in df["prompt"]:
        try:
            top_mean_ids, top_mean_vals, token_feature_ids = analyze_prompt_with_sae(
                prompt=prompt,
                model=model,
                sae=sae,
                hook_point=sae,
                topk_mean=10,
                topk_token=3,
            )
        except Exception as e:
            print(f"Error on prompt: {prompt[:50]}... -> {e}")
            top_mean_ids, top_mean_vals, token_feature_ids = None, None, None
    
        # all_top_mean_ids.append(top_mean_ids)
        # all_top_mean_vals.append(top_mean_vals)
        all_token_feature_ids.append(token_feature_ids)

    # print(len(all_token_feature_ids), all_token_feature_ids)
    # Save to DataFrame
    # SAE_ID = f"layer_{LAYER}/width_16k/canonical"
    # df[release+ "-layer" + str(layer) + "-" +"top_mean_ids"] = all_top_mean_ids
    # df[release+ "-layer" + str(layer) + "-" +"top_mean_vals"] = all_top_mean_vals
    df[release+ "-layer" + str(layer) + "-" +"token_feature_ids"] = all_token_feature_ids
    
    # save to CSV
    df.to_csv(OUTPUT_PATH, index=False, encoding="utf-8-sig", quoting=1)


# RELEASE + SAE_ID + TYPE
# LAYER = 25

LAYERS = [i for i in range(19, 26)]
# RELEASE = f"gemma-scope-2b-pt-{SAE_TYPE}-canonical"
SAE_TYPES = [
            "res", 
            # "mlp", 
            # "att"
            ]
DATASETS_NAMES = [
                # "emotion", 
                # "math", 
                # "mmlu", 
                "programming"
]

print(DATASETS_NAMES, SAE_TYPES)

for sae_type in SAE_TYPES:
    for layer in LAYERS:
        # load model 
        model_name = "gemma-2-2b"
        release = f"gemma-scope-2b-pt-{sae_type}-canonical"
        sae_id = f"layer_{layer}/width_16k/canonical"
        model, sae, hook_point = load_model_and_sae(model_name=model_name, release=release, sae_id=sae_id, device="cuda")
        print(sae_type, layer, hook_point)
        
        for DATASET_NAME in DATASETS_NAMES:
            DATASET_PATH = f"./{DATASET_NAME}.csv"
            OUTPUT_PATH =  f"./{DATASET_NAME}.csv"
            df = load_dataset(DATASET_PATH)
        
            # print("PROCESSING: ", DATASET_NAME)
        
            pipeline_process_datasets(df, DATASET_PATH, OUTPUT_PATH, model, sae, hook_point, release, layer)
        
        # free this SAE right away
        del sae
        torch.cuda.empty_cache()

    # free the model before moving on to the next SAE type
    del model
    torch.cuda.empty_cache()
